Move Repertories to Smanager
