+++ project/apps/bhs/models.py
@@ -1528,122 +1528,122 @@ class Chart(TimeStampedModel):
         return
 
 
-class Repertory(TimeStampedModel):
-    id = models.UUIDField(
-        primary_key=True,
-        default=uuid.uuid4,
-        editable=False,
-    )
+# class Repertory(TimeStampedModel):
+#     id = models.UUIDField(
+#         primary_key=True,
+#         default=uuid.uuid4,
+#         editable=False,
+#     )
 
-    STATUS = Choices(
-        (-10, 'inactive', 'Inactive',),
-        (0, 'new', 'New'),
-        (10, 'active', 'Active'),
-    )
+#     STATUS = Choices(
+#         (-10, 'inactive', 'Inactive',),
+#         (0, 'new', 'New'),
+#         (10, 'active', 'Active'),
+#     )
 
-    status = FSMIntegerField(
-        help_text="""DO NOT CHANGE MANUALLY unless correcting a mistake.  Use the buttons to change state.""",
-        choices=STATUS,
-        default=STATUS.active,
-    )
+#     status = FSMIntegerField(
+#         help_text="""DO NOT CHANGE MANUALLY unless correcting a mistake.  Use the buttons to change state.""",
+#         choices=STATUS,
+#         default=STATUS.active,
+#     )
 
-    # FKs
-    group = models.ForeignKey(
-        'Group',
-        related_name='repertories',
-        on_delete=models.CASCADE,
-    )
+#     # FKs
+#     group = models.ForeignKey(
+#         'Group',
+#         related_name='repertories',
+#         on_delete=models.CASCADE,
+#     )
 
-    chart = models.ForeignKey(
-        'Chart',
-        related_name='repertories',
-        on_delete=models.CASCADE,
-    )
+#     chart = models.ForeignKey(
+#         'Chart',
+#         related_name='repertories',
+#         on_delete=models.CASCADE,
+#     )
 
-    # Relations
-    statelogs = GenericRelation(
-        StateLog,
-        related_query_name='repertories',
-    )
+#     # Relations
+#     statelogs = GenericRelation(
+#         StateLog,
+#         related_query_name='repertories',
+#     )
 
-    # Internals
-    class Meta:
-        verbose_name_plural = 'repertories'
-        unique_together = (
-            ('group', 'chart',),
-        )
+#     # Internals
+#     class Meta:
+#         verbose_name_plural = 'repertories'
+#         unique_together = (
+#             ('group', 'chart',),
+#         )
 
-    class JSONAPIMeta:
-        resource_name = "repertory"
+#     class JSONAPIMeta:
+#         resource_name = "repertory"
 
-    def __str__(self):
-        return str(self.id)
+#     def __str__(self):
+#         return str(self.id)
 
-    # Permissions
-    @staticmethod
-    @allow_staff_or_superuser
-    @authenticated_users
-    def has_read_permission(request):
-        return True
+#     # Permissions
+#     @staticmethod
+#     @allow_staff_or_superuser
+#     @authenticated_users
+#     def has_read_permission(request):
+#         return True
 
-    @allow_staff_or_superuser
-    @authenticated_users
-    def has_object_read_permission(self, request):
-        roles = [
-            'SCJC',
-            'DRCJ',
-            'CA',
-            'Librarian',
-        ]
-        return any([
-            [item in roles for item in request.user.roles.values_list('name')],
-            # self.group.officers.filter(
-            #     person__user=request.user,
-            #     status__gt=0,
-            # ),
-        ])
+#     @allow_staff_or_superuser
+#     @authenticated_users
+#     def has_object_read_permission(self, request):
+#         roles = [
+#             'SCJC',
+#             'DRCJ',
+#             'CA',
+#             'Librarian',
+#         ]
+#         return any([
+#             [item in roles for item in request.user.roles.values_list('name')],
+#             # self.group.officers.filter(
+#             #     person__user=request.user,
+#             #     status__gt=0,
+#             # ),
+#         ])
 
-    @staticmethod
-    @allow_staff_or_superuser
-    @authenticated_users
-    def has_write_permission(request):
-        roles = [
-            'SCJC',
-            'DRCJ',
-            'Librarian',
-            'Manager',
-        ]
-        return any([item in roles for item in request.user.roles.values_list('name')])
+#     @staticmethod
+#     @allow_staff_or_superuser
+#     @authenticated_users
+#     def has_write_permission(request):
+#         roles = [
+#             'SCJC',
+#             'DRCJ',
+#             'Librarian',
+#             'Manager',
+#         ]
+#         return any([item in roles for item in request.user.roles.values_list('name')])
 
 
-    @allow_staff_or_superuser
-    @authenticated_users
-    def has_object_write_permission(self, request):
-        roles = [
-            'SCJC',
-            'DRCJ',
-            'Librarian',
-        ]
-        return any([
-            [item in roles for item in request.user.roles.values_list('name')],
-            # self.group.officers.filter(
-            #     person__user=request.user,
-            #     status__gt=0,
-            # ),
-        ])
+#     @allow_staff_or_superuser
+#     @authenticated_users
+#     def has_object_write_permission(self, request):
+#         roles = [
+#             'SCJC',
+#             'DRCJ',
+#             'Librarian',
+#         ]
+#         return any([
+#             [item in roles for item in request.user.roles.values_list('name')],
+#             # self.group.officers.filter(
+#             #     person__user=request.user,
+#             #     status__gt=0,
+#             # ),
+#         ])
 
-    # Transitions
-    @fsm_log_by
-    @transition(field=status, source='*', target=STATUS.active)
-    def activate(self, *args, **kwargs):
-        """Activate the Repertory."""
-        return
+#     # Transitions
+#     @fsm_log_by
+#     @transition(field=status, source='*', target=STATUS.active)
+#     def activate(self, *args, **kwargs):
+#         """Activate the Repertory."""
+#         return
 
-    @fsm_log_by
-    @transition(field=status, source='*', target=STATUS.inactive)
-    def deactivate(self, *args, **kwargs):
-        """Deactivate the Repertory."""
-        return
+#     @fsm_log_by
+#     @transition(field=status, source='*', target=STATUS.inactive)
+#     def deactivate(self, *args, **kwargs):
+#         """Deactivate the Repertory."""
+#         return
 
 
 class Award(TimeStampedModel):
