+++ project/apps/bhs/serializers.py
@@ -15,7 +15,6 @@ from .models import Person
 
 from .models import Award
 from .models import Chart
-from .models import Repertory
 
 
 class AwardSerializer(serializers.ModelSerializer):
@@ -57,7 +56,7 @@ class AwardSerializer(serializers.ModelSerializer):
 class GroupSerializer(serializers.ModelSerializer):
     permissions = DRYPermissionsField()
     included_serializers = {
-        'repertories': 'apps.bhs.serializers.RepertorySerializer',
+        # 'repertories': 'apps.bhs.serializers.RepertorySerializer',
         # 'members': 'apps.bhs.serializers.MemberSerializer',
         # 'officers': 'apps.bhs.serializers.OfficerSerializer',
     }
@@ -106,7 +105,7 @@ class GroupSerializer(serializers.ModelSerializer):
             'parent',
             # 'children',
 
-            'repertories',
+            # 'repertories',
             'permissions',
 
             'nomen',
@@ -120,7 +119,7 @@ class GroupSerializer(serializers.ModelSerializer):
 
     class JSONAPIMeta:
         included_resources = [
-            'repertories',
+            # 'repertories',
             # 'members',
             # 'officers',
         ]
@@ -241,7 +240,7 @@ class ChartSerializer(serializers.ModelSerializer):
     permissions = DRYPermissionsField()
 
     included_serializers = {
-        'repertories': 'apps.bhs.serializers.RepertorySerializer',
+        # 'repertories': 'apps.bhs.serializers.RepertorySerializer',
     }
 
     class Meta:
@@ -261,7 +260,7 @@ class ChartSerializer(serializers.ModelSerializer):
             'nomen',
             'image_id',
 
-            'repertories',
+            # 'repertories',
             'permissions',
         ]
         read_only_fields = [
@@ -271,26 +270,7 @@ class ChartSerializer(serializers.ModelSerializer):
 
     class JSONAPIMeta:
         included_resources = [
-            'repertories',
+            # 'repertories',
         ]
 
 
-class RepertorySerializer(serializers.ModelSerializer):
-    permissions = DRYPermissionsField()
-
-    class Meta:
-        model = Repertory
-        fields = [
-            'id',
-            'status',
-            'group',
-            'chart',
-            'permissions',
-        ]
-        validators = [
-            UniqueTogetherValidator(
-                queryset=Repertory.objects.all(),
-                fields=('group', 'chart'),
-                message='This chart already exists in your repertory.',
-            )
-        ]
