+++ project/apps/bhs/urls.py
@@ -9,7 +9,6 @@ from .views import GroupViewSet
 # from .views import OfficerViewSet
 from .views import PersonViewSet
 from .views import ChartViewSet
-from .views import RepertoryViewSet
 
 router = routers.DefaultRouter(
     trailing_slash=False,
@@ -21,5 +20,4 @@ router.register(r'group', GroupViewSet)
 router.register(r'award', AwardViewSet, basename='award')
 router.register(r'person', PersonViewSet)
 router.register(r'chart', ChartViewSet)
-router.register(r'repertory', RepertoryViewSet)
 urlpatterns = router.urls
