+++ project/apps/smanager/urls.py
@@ -8,6 +8,7 @@ from .views import ConventionViewSet
 from .views import ContestViewSet
 from .views import EntryViewSet
 from .views import SessionViewSet
+from .views import RepertoryViewSet
 
 router = routers.DefaultRouter(
     trailing_slash=False,
@@ -19,4 +20,5 @@ router.register(r'convention', ConventionViewSet)
 router.register(r'contest', ContestViewSet)
 router.register(r'entry', EntryViewSet)
 router.register(r'session', SessionViewSet)
+router.register(r'repertory', RepertoryViewSet)
 urlpatterns = router.urls
